Moved Runner definitions to a source file
